test(engine): separate Publish and Expire commands tests
